Fixed Nicolas Hack

+++ src/main/java/com/awards/service/VoteService.java
@@ -4,6 +4,7 @@ import com.awards.domain.Nominee;
 import com.awards.domain.Vote;
 import com.awards.repository.NomineeRepository;
 import com.awards.repository.VoteRepository;
+import com.awards.security.SecurityUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -23,15 +24,17 @@ public class VoteService {
     private NomineeRepository nomineeRepository;
 
     public void castVote(Vote vote){
+      if(vote.getUserId().equalsIgnoreCase(SecurityUtils.getCurrentLogin())){
         Vote existingVote = voteRepository.findExistingVote(vote.getUserId(), vote.getCategoryId());
         if(existingVote == null){
-            log.info("Casting a new vote {}", vote);
-            voteRepository.save(vote);
+          log.info("Casting a new vote {}", vote);
+          voteRepository.save(vote);
         }else{
-            log.info("Updating vote {}, new vote is {}", existingVote, vote);
-            voteRepository.delete(existingVote);
-            voteRepository.save(vote);
+          log.info("Updating vote {}, new vote is {}", existingVote, vote);
+          voteRepository.delete(existingVote);
+          voteRepository.save(vote);
         }
+      }
     }
 
     public List<Nominee> getResultsByCategory(String categoryId) {
